+++ pyviews/core/__init__.py
@@ -2,7 +2,7 @@
 
 from .error import ViewInfo, ViewsError
 from .binding import Binding, BindingError, BindingTarget
-from .node import Node, InstanceNode, Modifier, Property
+from .node import Node, InstanceNode, Modifier
 from .observable import Observable, ObservableEntity, InheritedDict
 from .reflection import import_path
 from .xml import XmlAttr, XmlNode