+++ pyviews/core/node.py
@@ -1,6 +1,5 @@
 """Core classes for creation from xml nodes"""
 
-from inspect import signature, Parameter
 from typing import Any, List, Callable
 
 from .binding import Binding
@@ -17,8 +16,7 @@ class Node:
         self._xml_node: XmlNode = xml_node
         self._globals: InheritedDict = InheritedDict() if node_globals is None else node_globals
         self._globals['node'] = self
-        self.attr_setter = _attr_setter
-        self.properties = {}
+        self.attr_setter = setattr
         self.on_destroy = lambda node: None
 
     @property
@@ -72,20 +70,13 @@ class Node:
         self._bindings = []
 
 
-def _attr_setter(node: Node, key, value):
-    if key in node.properties:
-        node.properties[key].set(value)
-    else:
-        setattr(node, key, value)
-
-
 class InstanceNode(Node):
     """Represents Node that wraps instance created from xml node"""
 
     def __init__(self, instance: Any, xml_node: XmlNode, node_globals: InheritedDict = None):
         super().__init__(xml_node, node_globals)
         self._instance = instance
-        self.attr_setter = _inst_attr_setter
+        self.attr_setter = _instance_attr_setter
 
     @property
     def instance(self):
@@ -93,39 +84,9 @@ class InstanceNode(Node):
         return self._instance
 
 
-def _inst_attr_setter(node: InstanceNode, key, value):
-    if key in node.properties:
-        node.properties[key].set(value)
-    else:
-        ent = node if hasattr(node, key) else node.instance
-        setattr(ent, key, value)
-
-
-class Property:
-    """Class to define property"""
-
-    def __init__(self, name, setter=None, node: Node = None):
-        self.name = name
-        self._value = None
-        self._setter = None
-        if setter:
-            args_count = len([p for p in signature(setter).parameters.values()
-                              if p.default == Parameter.empty])
-            self._setter = setter if args_count == 3 else \
-                lambda nd, value, previous: setter(nd, value)
-        self._node = node
-
-    def get(self):
-        """Returns value"""
-        return self._value
-
-    def set(self, value):
-        """Sets value"""
-        self._value = self._setter(self._node, value, self._value) if self._setter else value
-
-    def new(self, node: Node):
-        """Creates property for node"""
-        return Property(self.name, self._setter, node)
+def _instance_attr_setter(node: InstanceNode, key, value):
+    ent = node if hasattr(node, key) else node.instance
+    setattr(ent, key, value)
 
 
 Modifier = Callable[[Node, str, Any], None]