+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - create_node() function is made part of RenderingPipeline
 - RenderingPipeline is resolved by xml namespace and tag
 - Added RenderingIterator
+- get_inst_type renamed to get_type
+- create_inst renamed to create_instance
 - Rendering steps are renamed to pipes. Common pipes module is moved to core
 - Common modifiers module is moved to core
 - Removed CompiledExpression class
@@ -14,6 +16,7 @@
 - added execute() function to compilation package
 - removed injection of Expression class
 - CoreError is renamed to ViewsError
+- removed Node.properties and Property class
 
 ## 2.3.0